requirements.txt
# Core dependencies
google-cloud-bigquery>=3.3.5
tabulate>=0.9.0

# Web tools
requests>=2.28.1
beautifulsoup4>=4.11.1
aiohttp>=3.8.3
playwright>=1.37.0

# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.59.8 # o1 support
anthropic>=0.42.0
python-dotenv>=1.0.0
google-generativeai

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.70.0
